backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

